+++ www/admin/assets/js/jscalendar/lang/calendar-fi.js
@@ -85,8 +85,6 @@ Calendar._TT["NEXT_YEAR"] = "Seur. vuosi (paina hetki, näet valikon)";
 Calendar._TT["SEL_DATE"] = "Valitse päivämäärä";
 Calendar._TT["DRAG_TO_MOVE"] = "Siirrä kalenterin paikkaa";
 Calendar._TT["PART_TODAY"] = " (tänään)";
-Calendar._TT["MON_FIRST"] = "Näytä maanantai ensimmäisenä";
-Calendar._TT["SUN_FIRST"] = "Näytä sunnuntai ensimmäisenä";
 Calendar._TT["CLOSE"] = "Sulje";
 Calendar._TT["TODAY"] = "Tänään";
 Calendar._TT["TIME_PART"] = "(Shift-) Klikkaa tai liikuta muuttaaksesi aikaa";
@@ -96,3 +94,6 @@ Calendar._TT["DEF_DATE_FORMAT"] = "%d.%m.%Y";
 Calendar._TT["TT_DATE_FORMAT"] = "%d.%m.%Y";
 
 Calendar._TT["WK"] = "Vko";
+
+Calendar._TT["WEEKEND"] = "0,6";
+Calendar._TT["DAY_FIRST"] = "Näytä %s ensimmäisenä";