+++ www/admin/assets/js/jscalendar/lang/calendar-jp.js
@@ -33,8 +33,6 @@ Calendar._TT["NEXT_YEAR"] = "翌年";
 Calendar._TT["SEL_DATE"] = "日付選択";
 Calendar._TT["DRAG_TO_MOVE"] = "ウィンドウの移動";
 Calendar._TT["PART_TODAY"] = " (今日)";
-Calendar._TT["MON_FIRST"] = "月曜日を先頭に";
-Calendar._TT["SUN_FIRST"] = "日曜日を先頭に";
 Calendar._TT["CLOSE"] = "閉じる";
 Calendar._TT["TODAY"] = "今日";
 
@@ -43,3 +41,6 @@ Calendar._TT["DEF_DATE_FORMAT"] = "y-mm-dd";
 Calendar._TT["TT_DATE_FORMAT"] = "%m月 %d日 (%a)";
 
 Calendar._TT["WK"] = "週";
+
+Calendar._TT["WEEKEND"] = "0,6";
+Calendar._TT["DAY_FIRST"] = "%s曜日を先頭に";