+++ www/admin/assets/js/jscalendar/lang/calendar-ko.js
@@ -107,8 +107,6 @@ Calendar._TT["NEXT_YEAR"] = "다음 해 (길게 누르면 목록)";
 Calendar._TT["SEL_DATE"] = "날짜를 선택하세요";
 Calendar._TT["DRAG_TO_MOVE"] = "마우스 드래그로 이동 하세요";
 Calendar._TT["PART_TODAY"] = " (오늘)";
-Calendar._TT["MON_FIRST"] = "월요일을 한 주의 시작 요일로";
-Calendar._TT["SUN_FIRST"] = "일요일을 한 주의 시작 요일로";
 Calendar._TT["CLOSE"] = "닫기";
 Calendar._TT["TODAY"] = "오늘";
 Calendar._TT["TIME_PART"] = "(Shift-)클릭 또는 드래그 하세요";
@@ -118,3 +116,6 @@ Calendar._TT["DEF_DATE_FORMAT"] = "%Y-%m-%d";
 Calendar._TT["TT_DATE_FORMAT"] = "%b/%e [%a]";
 
 Calendar._TT["WK"] = "주";
+
+Calendar._TT["WEEKEND"] = "0,6";
+Calendar._TT["DAY_FIRST"] = "%s을 한 주의 시작 요일로";