+++ www/admin/assets/js/jscalendar/lang/calendar-lt.js
@@ -101,8 +101,6 @@ Calendar._TT["NEXT_YEAR"] = "Kiti metai (laikykite, jei norite meniu)";
 Calendar._TT["SEL_DATE"] = "Pasirinkite datą";
 Calendar._TT["DRAG_TO_MOVE"] = "Tempkite";
 Calendar._TT["PART_TODAY"] = " (šiandien)";
-Calendar._TT["MON_FIRST"] = "Pirma savaitės diena - pirmadienis";
-Calendar._TT["SUN_FIRST"] = "Pirma savaitės diena - sekmadienis";
 Calendar._TT["CLOSE"] = "Uždaryti";
 Calendar._TT["TODAY"] = "Šiandien";
 Calendar._TT["TIME_PART"] = "Spustelkite arba tempkite jei norite pakeisti";
@@ -112,3 +110,6 @@ Calendar._TT["DEF_DATE_FORMAT"] = "%Y-%m-%d";
 Calendar._TT["TT_DATE_FORMAT"] = "%A, %Y-%m-%d";
 
 Calendar._TT["WK"] = "sav";
+
+Calendar._TT["WEEKEND"] = "0,6";
+Calendar._TT["DAY_FIRST"] = "Pirma savaitės diena - %s";