+++ www/admin/assets/js/jscalendar/lang/calendar-sv.js
@@ -80,8 +80,6 @@ Calendar._TT["NEXT_YEAR"] = "Följande år (håll för menu)";
 Calendar._TT["SEL_DATE"] = "Välj datum";
 Calendar._TT["DRAG_TO_MOVE"] = "Drag för att flytta";
 Calendar._TT["PART_TODAY"] = " (idag)";
-Calendar._TT["MON_FIRST"] = "Visa måndag först";
-Calendar._TT["SUN_FIRST"] = "Visa söndag först";
 Calendar._TT["CLOSE"] = "Stäng";
 Calendar._TT["TODAY"] = "Idag";
 Calendar._TT["TIME_PART"] = "(Skift-)klicka eller drag för att ändra tid";
@@ -91,3 +89,6 @@ Calendar._TT["DEF_DATE_FORMAT"] = "%Y-%m-%d";
 Calendar._TT["TT_DATE_FORMAT"] = "%A %d %b %Y";
 
 Calendar._TT["WK"] = "vecka";
+
+Calendar._TT["WEEKEND"] = "0,6";
+Calendar._TT["DAY_FIRST"] = "Visa %s först";