+++ www/admin/assets/js/jscalendar/lang/calendar-tr.js
@@ -46,8 +46,6 @@ Calendar._TT["NEXT_YEAR"] = "Sonraki Yýl (Menü için basýlý tutunuz)";
 Calendar._TT["SEL_DATE"] = "Tarih seçiniz";
 Calendar._TT["DRAG_TO_MOVE"] = "Taþýmak için sürükleyiniz";
 Calendar._TT["PART_TODAY"] = " (bugün)";
-Calendar._TT["MON_FIRST"] = "Takvim Pazartesi gününden baþlasýn";
-Calendar._TT["SUN_FIRST"] = "Takvim Pazar gününden baþlasýn";
 Calendar._TT["CLOSE"] = "Kapat";
 Calendar._TT["TODAY"] = "Bugün";
 
@@ -56,3 +54,7 @@ Calendar._TT["DEF_DATE_FORMAT"] = "dd-mm-y";
 Calendar._TT["TT_DATE_FORMAT"] = "d MM y, DD";
 
 Calendar._TT["WK"] = "Hafta";
+
+Calendar._TT["WEEKEND"] = "0,6";
+Calendar._TT["DAY_FIRST"] = "Takvim %s gününden baþlasýn";
+